+++ NEWS.md
@@ -1,3 +1,24 @@
+# 2.0.10.9003
+
+* Added two internal functions `GetDistanceInfo` and `GetExposureInfo` to
+retrieve information from output returned by `MAPAssignActivity`
+
+# 2.0.10.9002
+
+* Added signature etiologies information to the reconstructed spectrum plot when calling `MAPAssignActivity`
+
+# 2.0.10.9001
+
+* Added `LLHSpectrumMAP` and tests
+
+# 2.0.10
+
+* Added `MAPAssignActivity` and tests
+
+# 2.0.9.9000
+
+* Added sample information to show to the user when calling `MAPAssignActivity1`
+
 # 2.0.8.9005
 
 * Round the non integers in spectrum to integers and give a warning when calling